requirements.txt
asgiref==3.2.10
Brotli==1.0.9
dj-database-url==0.5.0
Django==3.1.2
django-crispy-forms==1.9.2
django-environ==0.4.5
django-heroku==0.3.1
django-materializecss-form==1.1.17
environ==1.0
gunicorn==20.0.4
Pillow==8.0.0
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.4.0
whitenoise==5.2.0
